Refactor AllPrivilegesPermittedShardingSpherePrivileges
