+++ shardingsphere-kernel/shardingsphere-authority/shardingsphere-authority-core/src/main/java/org/apache/shardingsphere/authority/provider/natived/model/privilege/NativePrivileges.java
@@ -32,7 +32,7 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
- * Native Privileges.
+ * Native privileges.
  */
 @Getter
 @EqualsAndHashCode
