+++ shardingsphere-kernel/shardingsphere-authority/shardingsphere-authority-core/src/main/java/org/apache/shardingsphere/authority/provider/schema/model/privilege/SchemaPrivilegesPermittedShardingSpherePrivileges.java
@@ -27,7 +27,7 @@ import java.util.Collection;
 import java.util.Set;
 
 /**
- * Schema privilegesPermittedShardingSpherePrivileges.
+ * Schema permitted privileges.
  */
 @RequiredArgsConstructor
 public final class SchemaPrivilegesPermittedShardingSpherePrivileges implements ShardingSpherePrivileges {
