+++ shardingsphere-kernel/shardingsphere-authority/shardingsphere-authority-core/src/main/java/org/apache/shardingsphere/authority/provider/simple/AllPrivilegesPermittedAuthorityProviderAlgorithm.java
@@ -18,10 +18,7 @@
 package org.apache.shardingsphere.authority.provider.simple;
 
 import lombok.Getter;
-import org.apache.shardingsphere.authority.model.AccessSubject;
 import org.apache.shardingsphere.authority.model.AuthorityRegistry;
-import org.apache.shardingsphere.authority.model.PrivilegeType;
-import org.apache.shardingsphere.authority.model.ShardingSpherePrivileges;
 import org.apache.shardingsphere.authority.registry.AllPermittedAuthorityRegistry;
 import org.apache.shardingsphere.authority.spi.AuthorityProviderAlgorithm;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
@@ -36,8 +33,6 @@ import java.util.Properties;
  */
 @Getter
 public final class AllPrivilegesPermittedAuthorityProviderAlgorithm implements AuthorityProviderAlgorithm {
-    
-    private static final ShardingSpherePrivileges INSTANCE = new AllPrivilegesPermittedShardingSpherePrivileges();
     
     private Properties props;
     
@@ -54,27 +49,5 @@ public final class AllPrivilegesPermittedAuthorityProviderAlgorithm implements A
     @Override
     public String getType() {
         return "ALL_PRIVILEGES_PERMITTED";
-    }
-    
-    public static final class AllPrivilegesPermittedShardingSpherePrivileges implements ShardingSpherePrivileges {
-        
-        @Override
-        public void setSuperPrivilege() {
-        }
-        
-        @Override
-        public boolean hasPrivileges(final String schema) {
-            return true;
-        }
-        
-        @Override
-        public boolean hasPrivileges(final Collection<PrivilegeType> privileges) {
-            return true;
-        }
-        
-        @Override
-        public boolean hasPrivileges(final AccessSubject accessSubject, final Collection<PrivilegeType> privileges) {
-            return true;
-        }
     }
 }
